Finish basic EquipAction functionality
If an item slot is blocked, it is automatically unequipped
Added childAction-functionality to UnequipAction to allow adjusting its functionality so that it works as a sub-process in EquipAction. Might generalize this into a new Behavior ChildActionableBehavior or so later on.
Define Tools->getTransaction() to prevent different game actions defining several transactions at the same time, which results in exceptions being thrown all around
Fix wrong equipment slot names in EquipmentWidget view file for good